fix: convert story point to signed integer (#2394)
* fix: convert story point to signed integer

* test: modify e2e test to cover the case of minus story point

* fix: remove debug code

+++ plugins/tapd/tasks/story_converter.go
@@ -67,7 +67,7 @@ func ConvertStory(taskCtx core.SubTaskContext) error {
 				Title:                toolL.Name,
 				Type:                 toolL.StdType,
 				Status:               toolL.StdStatus,
-				StoryPoint:           uint(toolL.Size),
+				StoryPoint:           int64(toolL.Size),
 				OriginalStatus:       toolL.Status,
 				ResolutionDate:       (*time.Time)(toolL.Completed),
 				CreatedDate:          (*time.Time)(toolL.Created),
